server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Student Personalizer API Server"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sqlmodel>=0.0.22",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "python-multipart>=0.0.9",
    "psycopg[binary]>=3.2.0",
    "httpx>=0.27.0",
    "numpy>=2.4.2",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "joblib>=1.5.3",
    "scikit-learn>=1.8.0",
    "python-dotenv>=1.2.1",
    "psycopg2-binary>=2.9.11",
    "argon2-cffi>=23.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-json-logger>=3.3.0",
    "alembic>=1.14.0",
    "cryptography>=44.0.0",
    "pydantic-settings>=2.0",
]

[project.scripts]
server = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = ["ruff>=0.9.0", "pytest>=8.0.0"]

[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
